requirements.txt
aiofiles==23.2.1
annotated-types==0.6.0
anyio==4.3.0
attrs==19.3.0
Automat==0.8.0
bcrypt==4.1.2
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloud-init==23.4.4
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
DateTime==5.5
dbus-python==1.2.16
distro==1.4.0
distro-info==0.23+ubuntu1.1
dnspython==2.6.1
ecdsa==0.18.0
email_validator==2.1.1
entrypoints==0.3
fastapi==0.110.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httplib2==0.14.0
httptools==0.6.1
httpx==0.27.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==2.1.5
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.26.4
oauthlib==3.1.0
orjson==3.9.15
packaging==24.0
pandas==2.2.1
passlib==1.7.4
pexpect==4.6.0
pluggy==1.4.0
psycopg2-binary==2.9.9
pyasn1==0.4.2
pyasn1-modules==0.2.1
pydantic==2.6.4
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyMySQL==1.1.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyrsistent==0.15.5
pyserial==3.4
pytest==8.1.1
pytest-asyncio==0.23.6
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.9.0.post0
python-debian==0.1.36+ubuntu1.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
rsa==4.9
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sniffio==1.3.1
sos==4.5.6
SQLAlchemy==2.0.29
ssh-import-id==5.10
starlette==0.36.3
systemd-python==234
Twisted==18.9.0
typing_extensions==4.10.0
tzdata==2024.1
ubuntu-pro-client==8001
ufw==0.36
ujson==5.9.0
unattended-upgrades==0.1
urllib3==1.25.8
uvicorn==0.29.0
uvloop==0.19.0
wadllib==1.3.3
watchfiles==0.21.0
websockets==12.0
zipp==1.0.0
zope.interface==4.7.1
